+++ feature-deep-linking/src/main/java/io/novafoundation/nova/feature_deep_linking/di/DeepLinkingFeatureApi.kt
@@ -1,6 +1,6 @@
 package io.novafoundation.nova.feature_deep_linking.di
 
-import io.novafoundation.nova.app.root.presentation.deepLinks.handlers.AssetDetailsDeepLinkHandler
+import io.novafoundation.nova.feature_deep_linking.presentation.handling.handlers.AssetDetailsDeepLinkHandler
 import io.novafoundation.nova.feature_deep_linking.presentation.deferred.ReferralInstallHandler
 import io.novafoundation.nova.feature_deep_linking.presentation.handling.handlers.ReferendumDeepLinkHandler
 import io.novafoundation.nova.feature_deep_linking.presentation.handling.RootDeepLinkHandler
