+++ feature-push-notifications/src/main/java/io/novafoundation/nova/feature_push_notifications/presentation/handling/NotificationHandlerExt.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import androidx.core.app.NotificationCompat
-import io.novafoundation.nova.app.root.presentation.deepLinks.handlers.AssetDetailsDeepLinkData
+import io.novafoundation.nova.feature_deep_linking.presentation.handling.handlers.AssetDetailsDeepLinkData
 import io.novafoundation.nova.feature_deep_linking.presentation.handling.handlers.ReferendumDeepLinkData
 import io.novafoundation.nova.common.utils.asGsonParsedNumber
 import io.novafoundation.nova.feature_account_api.domain.interfaces.AccountRepository
